python/packages/corlinman-server/pyproject.toml
[project]
name = "corlinman-server"
version = "0.1.0"
description = "grpc.aio server entrypoint for the corlinman Python AI plane."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "corlinman contributors" }]

dependencies = [
    "grpcio>=1.67",
    "structlog>=25.1",
    "python-json-logger>=3.2",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp>=1.27",
    "corlinman-agent",
    "corlinman-embedding",
    "corlinman-providers",
    "corlinman-grpc",
]

[tool.uv.sources]
corlinman-agent = { workspace = true }
corlinman-embedding = { workspace = true }
corlinman-providers = { workspace = true }
corlinman-grpc = { workspace = true }

[project.scripts]
corlinman-python-server = "corlinman_server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corlinman_server"]
